satisfy: add inference context to the Satisfier trait
This is a bit annoying for users of the Satisfier trait: they need to
construct a type inference and keep it around for the satisfaction logic
to use. If the satisfier supports assembly this is necessary.

If it doesn't support assembly, which is probably the more common case,
then there's no need for this ... but because of lack of specialization
and a general lack of expressivity in Rust's type system, they gotta do
it anyway.