types: introduce helper type to hold token alongside slab
In the next couple commits we will replace the mutexes in the union-bound
algorithm with GhostCells. Happily, we can reuse the existing GhostToken
in types::Context as an access-control token.

However, to do this we need to refactor a couple algorithms so that we
can fish the token recursively through unify() and bind(), which naively
would require the union-bound algorithm to know about our ContextInner
type. To avoid this abstraction violation, we introduce the generic
WithContextToken wrapper, so the union-bound algorithm can work with "a
token and some arbitrary data" without knowing or understanding anything
about the arbitrary data.