types: replace LockedContext with a mutex guard around ContextInner
Now that we have the dedicated ContextInner type, we don't need a
dedicated LockedContext type. It's also conceptually simpler to have
a "Context, which mutex-wraps a ContextInner, which implements all
the real methods" rather than having this auxiliary "locked context"
type which isn't directly related to Context.

(And ofc there is WithGhostToken in the middle, but the reader of
the code can hopefully mostly ignore this.)

This gets rid of a lot of `.inner`s which is nice.