ffi: don't flatten CRawInput structure
I can't find any citation that this is safe to do. We don't actually use
this structure in our Rust code; we populate it by calling a c_set function
which is defined on the C side, and then pass it to C functions. So no code
needs to be changed.

However, I want to remove the c_set_* functions because they hide life
information (and in fact, are unsound, as we currently use them) and serve
to obfuscate code. They seem to have come from #48 #49 #50 #51 in which
Sanket seems to have copied the corresponding Haskell code into Rust, where
it is unnecessary and dangerous.