ffi: replace some "pointer to assumed length" types with optional references
This adds some lifetimes to our raw C structures, which in the next commits
will be used to enforce that we are not giving the C code pointers to local
variables or other unsound things. (In fact, our existing code does do this.)
As a starting point, replace a few "easy" FFI fields with stronger types.

We do not change the pointers to value types, which are listed in Rust as
"NULL, or a pointer to 9 bytes, or to 33" because it was not clear that
there was any better representation for this than *const u8.

Through the C FFI we have structs which have listed invariants of the form
"either this is NULL or it points to an array of 33 bytes". We further have
a guarantee (because we are maintainers of both the C and Rust code, and we
are responsible citizens) that these const pointers will never be mutated
through by the C code.

The rest of this comment is to justify this change:

According to the nomicon, we may use references in FFI types as long as we
promise they won't be NULL and that we won't violate the aliasing rules
(which for const pointers, simply means we won't cast them to non-const
pointers):

https://doc.rust-lang.org/nomicon/ffi.html#interoperability-with-foreign-code

The repr(C) docs further say that if our pointers might be NULL, we can
use Option<&T>:

https://doc.rust-lang.org/nomicon/other-reprs.html#reprc

Finally, where the C code takes a pointer to a u8 with an implicit claim
that this is actually a pointer to 33 contiguous u8s, we may use pointers
or references to Rust arrays, which is permissible because:

* all thin pointers have the same size and representation, justified by
  the Rust reference https://doc.rust-lang.org/reference/expressions/operator-expr.html#r-expr.as.pointer.sized
* arrays are guaranteed to be contiguous in memory in both C and Rust