ffi: delete the c_set_rawElementsOutput method
(The following commits will delete the rest of these.) The role of these
functions is to marshall Rust data into C structures. This is implemented
in env.c, on the C side of the FFI boundary, and use raw pointers and
out-pointers. None of this is necessary because Rust has a well-defined
FFI interface with C.

Furthermore, it serves to obfuscate the lifetimes and provenance of the
various pointers at play.

We can remove this function, replace its call sites with direct construction
of the Rust version of the C structure, and replace raw pointers (where
possible) with references that have named lifetimes. Then our Rust code
is formally safe and the compiler will give us much more help.