ffi: don't use empty enums for opaque types
The nomicon doesn't like this. From the bottom of

https://doc.rust-lang.org/nomicon/ffi.html

"Notice that it is a really bad idea to use an empty enum as FFI type. The
compiler relies on empty enums being uninhabited, so handling values of
type &Empty is a huge footgun and can lead to buggy program behavior (by
triggering undefined behavior)."

They give an example that uses a phantom to also remove Send, Sync, and
Unpin. I did not do this. I think all these traits are fine, because the
type has no self-references and is never mutated by the C code except
when it is freed, and then only through a `*mut` pointer.