bit_machine: return concrete type from `as_bit_iter`
Returning a concrete type lets us get more trait methods rather than just
`Iterator`. In particular we should be able to get `ExactSizeIterator`
and `Fuse` which are important for efficiency.

We also want to name this type elsewhere, and even the existential one
was annoying to type, so we add an alias for it.