bit_machine: replace `ExecTracker` API with a much more general one
This replaces the `ExecTracker` API with one which is able to detect
every node, not just cases and jets; which is able to read the input
value for every node (as a bit iterator which can be converted to a
value with Value::from_padded_bits) and the output value for terminal
nodes; and which can do all the existing things that the tracker can
do.

I suspect we want to add some examples or unit tests, in particular
around "debug nodes".

Fixes #324