bit_machine: rename private new/move/drop methods and variants
There are three basic operations on the frame stack in the bit machine:
creating a new write frame, popping the write stack onto the read stack,
and dropping a read frame.

Our code uses shorthand like "new_frame" and "drop_frame" which makes
it easy to forget which stacks these operations are operating on. This
makes it harder to understand the code for somebody who doesn't currently
have the bit machine algorithm loaded into their head.

Rename to make stuff easier to follow.