value: display left and rights as L(x) and R(x); drop DispUnlessUnit
Previously we would display left branches as 0 and right branches as 1,
and had a special case where if the child node was unit, we would not
display the inner (ε) to reduce noise.

Now that we special-case all bitstrings, in particular bitstrings of
length 1, this special-case is unnecessary and actually confusing. Now
if we encounter a sum type we know it's *not* a bit and it would be
clearer to indicate it as L or R, and if the child is a unit, we should
unconditionally display that.